requirements.txt
astroid==1.4.9
autopep8==1.2.4
Django==1.10.3
django-bootstrap3==8.0.0
django-extensions==1.7.5
djangorestframework==3.5.3
Faker==0.7.3
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.3
mysql-connector==2.1.4
olefile==0.44
pep8==1.7.0
Pillow==4.0.0
pkg-resources==0.0.0
pylint==1.6.4
python-dateutil==2.6.0
six==1.10.0
wrapt==1.10.8
yapf==0.15.0
